Separated User class from main.cpp
