+++ LibraryManager/LibraryManager.vcxproj
@@ -21,12 +21,14 @@
   <ItemGroup>
     <ClCompile Include="src\Classes\BookClass.cpp" />
     <ClCompile Include="src\Classes\LibraryClass.cpp" />
+    <ClCompile Include="src\Classes\UserClass.cpp" />
     <ClCompile Include="src\main.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="src\BookClass.h" />
     <ClInclude Include="src\Classes\BookClass.h" />
     <ClInclude Include="src\Classes\LibraryClass.h" />
+    <ClInclude Include="src\Classes\UserClass.h" />
   </ItemGroup>
   <PropertyGroup Label="Globals">
     <VCProjectVersion>17.0</VCProjectVersion>
